policy: Fix flakiness of policy.Printers test
The test was flaky, likely because of a race between the print
preview closing and the print management app opening. When the print
preview closed after the print manangement app had already opened, the
browser would steal focus from the print manangement app. To avoid
this race condition, we now wait for the print preview dialog to close
before attempting to open the print preview app.

BUG=b:220977394
TEST=Ran test multiple times to verify that the print manangement app no
longer opens in the background.

Change-Id: Ib299f41c5c19ba80fb8d7b13f68186e6fa459e1a