tast-tests: stabilize crostini.AppGeditInstallUninstall.* tests
crostini.AppGeditInstallUninstall.* tests have been failing,
because they looked for a gedit icon in the toolbar. However,
on tablets the toolbar is closed by default. As result,
the tests timed out without finding the icon.

Do not rely on toolbar to interact with gedit: use
a11y tree and keyboard shortcuts instead.

BUG=b:232369175
TEST=tast run dut crostini.AppGeditInstallUninstall.bullseye_stable

Change-Id: I97da84ab6c6fc0a045d5f8aa7a4fd376253b0b2b