network: VPNUI: Add l2tp_ipsec_cert test
This patch adds the l2tp_ipsec_cert test to VPNUI. Since this test needs
the netcertstore setup, also extends the vpnResetShill fixture to
support this scenario. The new added fixture will do an additional
step to start the Chrome on top of the original one. Note that we cannot
use the existing chromeLoggedIn fixture since Chrome setup has to happen
after the vpnShillReset setup, this can be done by make vpnShillReset
fixture as the parent of the former fixture, but it seems that we cannot
return the value from vpnShillReset fixture in this way.

BUG=b:224875028
TEST=tast run $DUT network.VPNUI.*
TEST=tast run $DUT network.VPNConnect.*

Change-Id: I2ad9f013b496832711ffb2c37876a5cd75a847d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3645951
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>