webrtc: instantiate grid videos after establishing the PeerConnection
Oftentimes while WebRTC establishes a PeerConnection it creates 
more HW video decoders than actually needed (e.g. because it probes
for hw support for codec profiles etc). These decoders are transient 
but they use up instances needed for the ancillary <video> feeds,
causing the test to be flaky. 

This CL changes the logic so that the PeerConnection is established
first (at that point the PeerConnection codec must be fully resolved) 
before launching the ancillary video playbacks.

BUG=b:231982835
TEST=webrtc.RTCPeerConnectionPerf.vp8_hw_multi_vp9_4x4

Change-Id: I31692b18e3d261f06d4f216cc46713142a8a4d6d