hwsec: Fix USS flag file creation in PIN test
Fix the hwsec.PinWeaverWithAuthFactor test to correctly create the USS
flag file on the disk.

Before this commit, neither the file creation nor the mkdir command
succeeded, but the error was swallowed by the tast test. This led to
test failures: the test was attempting to run AddAuthFactor with a PIN
factor on a non-USS cryptohome, which isn't supported at the moment
(causing "CRYPTOHOME_ERROR_NOT_IMPLEMENTED"==4 errors).

Also remove a no-op and anyway wrong "rm low_entropy_creds" command
during the user setup.

BUG=b:228572025
TEST=tast run $DUT hwsec.PinWeaverWithAuthFactor

Change-Id: Ie7e48bdc7e1e412762edab99f4cd0d2fc02ec3ef