Tast test for launcher keyboard shortcut search
Adds test cases that search for queries expected to match a keyboard
shortcut, and verify that search returns the associated shortcut. For a
subset of test cases, verifies that the keyboard shortcut is the first
returned result. Not testing the category explicitly, as the result
could be returned as an answer card, best match, or a help category
result.

BUG=None
TEST=tast run launcher.ShortcutSearch*
TEST=tast unit test passed

Change-Id: I15de962e7016576537a5995fc3a2bc9937d75548
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3645663
Reviewed-by: Rachel Wong <wrong@chromium.org>
Reviewed-by: Yulun Wu <yulunwu@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Tested-by: Toni Barzic <tbarzic@chromium.org>