tast-tests: stabilize crostini fixtures
crostini fixtures setup has a separate step to uninstall csrostini
if it happens to be installed at the start of the fixture. This
reduces state sharing and guarantees, that we will not have crostini
installed before we attempt to install it in the fixture. Unfortunately
this code has a typo. As a result we skip this check 100% of time.

Fix a typo. This will enable checking for crostini presence before the
test and remove it, if found installed.

BUG=b:231630463
TEST=tast run dut crostini.AppGeditInstallUninstall.buster_stable

Change-Id: I494f47b9d039f36a7eab429b4bd4ea3ce0415adb