tast-tests: stabilize crostini fixtures
crostini fixtures setup has a separate step to uninstall csrostini
if it happens to be installed at the start of the fixture. This
reduces state sharing and guarantees, that we will not have crostini
installed before we attempt to install it in the fixture. Unfortunately
this code has a typo. As a result we skip this check 100% of time.

Fix a typo. This will enable checking for crostini presence before the
test and remove it, if found installed. Introduced cleanup context for
deferred cleanup actions.

BUG=b:231630463
TEST=tast run dut crostini.AppGeditInstallUninstall.buster_stable

Change-Id: I494f47b9d039f36a7eab429b4bd4ea3ce0415adb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3645830
Commit-Queue: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: David Munro <davidmunro@google.com>