tast-tests: Do not check EAP.ClientCertType in rollback
We need to verify the preservation of the networks after rollback.
Because of b/227740677 the test includes a work around to make sure
the value is not empty. However, if this bug is fixed, because of
b/227734735, the tests fail.

Remove the workaround check to avoid these failures until both bugs
have been fixed.

BUG=b:227740677, b:227734735
TEST=tast -verbose run localhost:2233 policy.EnterpriseRollbackInPlace

Change-Id: I3ad4d13b24e0a16ed155204c94c5cc15b819999d