telemetryextension: Fix organization of jacuzzi board
CL crrev/c/3565639 specified that it would remove the jacuzzi
board from the target list, but actually removed it from
the allow list. This commit fixes that by adding the jacuzzi
board back to the allow list and remove it from the target
list.

BUG=b:203402917
TEST=tast run <IP> telemetryextension.Compliance.*

Change-Id: Ib81596c8d29fd48355a6887c18089a299b1be3d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3647290
Reviewed-by: Mahmoud Aly <mgawad@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Tested-by: Bastian Kersting <bkersting@google.com>
Commit-Queue: Bastian Kersting <bkersting@google.com>