arc/gio: Refactor GIO launch test
With the addition of the educational dialog, there needs to be an additional click at the beginning of the action sequence to click through the input overlay educational dialog pop-up. In addition, due to b/231097154, the educational dialog screen causes uiauto to fail, so the entire launch detection test must be refactored to use ACUITI instead.

BUG=b:229156145
TEST=tast run $DUT arc.InputOverlayLaunch*
TEST=tast unit test passed.

Change-Id: I5cc22a91dcb19c530a9c4db0bd9bcdb208f335f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3646132
Reviewed-by: Cici Ruan <cuicuiruan@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Joseph Pang <pjlee@google.com>
Tested-by: Joseph Pang <pjlee@google.com>