wifi: Refactor test fixture to fully support multiple peers
Refactors TestFixture's internals to simplify handling of multiple DUTs.
Adds support for multiple DUTs in various routines, including
WiFi connection/disconnection. The support is either backwards-compatible
(when function was used in multiple files), or as drop-in replacement
(when function was used 1-3 times).
This way we can setup multiple DUTs in a simple for loop.

BUG=b:225176836
TEST=tast run -companiondut=cd1:$PEER_HOST $DUT wifi.TDLS

Change-Id: I0aafb19167ebd4f8c8786013d904f57531e9b743