tast-tests: Reproduce problem
I do not intend to land this CL. It is just for help with b/244356321.
It reproduces the underlying problem with as little code as possible.

The point here is that the drag should move the window, but it only
restores the window from maximized to normal state. It looks like this:
https://photos.app.goo.gl/oG6ocrdbbfFKFLAXA

BUG=b:244356321
TEST=tast run <DUT> ui.DebugBug1324662

Change-Id: I505a4b4f0f78f0f298f1bf739572711e7affeae4