policy: Increase tolerance on time values to avoid flakes
We have seen some bots being slow enough to have time values off by 4-5
seconds. Increasing tolerance to 7 seconds should be sufficient to avoid
flakes. We also increase the delay to avoid overlapping values.

This also adds a wait and page refresh after clicking "Reload policies"
in Lacros. Unlike in Ash browser, policies are not guaranteed to be
reloaded when button becomes enabled again and page needs to be
reloaded.

BUG=chromium:1325837,chromium:1326565
TEST=ran on local DUT, passed 10 times

Change-Id: I79077086bdbf633544b7f6b5d8220033dc7e9258