tast-tests: Refactor legacy bridge and veth support to common
This change refactors the logic around managing bridges and
veths out of the legacy router implementation into the common
package so that it may be reused in other router
implementations, most notably OpenWrt.

BUG=b:211517577
TEST=Successfully ran tast test wifi.RoamFT.psk, which uses the
bridge and veth support, on a wificell with a legacy router.

Change-Id: Ieafa5492ef5b848d3281fe178192bceb0c29c128
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3589080
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>