tast-tests: Add OpenWrt Bridge, Veth, and VethBridgeBinding support
This change updates the common veth support to allow for
automatically resolving the iface name of a vethPeer,
which is necessary on OpenWrt as it the name cannot be
set like it can on Legacy routers. Since Legacy routers
do not support for iface alias name resolution like OpenWrt
routers do, the veth peer names are not resolved for
Legacy routers. In addition, OpenWrt support for the Bridge,
Veth, and VethBridgeBinding router support interfaces has
been added utilizing the common package functions for bridge
and veth.

BUG=b:211517577
TEST=Successfully ran tast test wifi.RoamFT.psk, which uses the
bridge and veth support, on a wificell with a legacy router and
a wificell with an OpenWrt router.

Change-Id: If14c88e29a63dd3a671d0d82c82a3b7f722e634b