(Not for Submission) Test CL that shows drag issue in overview mode
When creating two windows in lacros and entering overview mode,
the first window is dragged / snapped to the left successfully,
but the second window fails to be dragged. No error is actually
thrown, and the cursor moves as if it actually dragged the window.

Unclear if this bug has the same root cause as the one in
chromium:1324662, but they appear to be both failures in dragging
a lacros window properly.

Video showing test result:
https://photos.app.goo.gl/Gz1C4QDAAWSE7LDj7

BUG=chromium:1324662
TEST=tast run <DUT> ui.DebugBug1324662

Change-Id: Ib6470029a6247aab31e32303ec84924f854763d9