tast-tests: Update ip LinkWithPrefix  to use full 'ip link show' format
The LinkWithPrefix function relied on the brief output format for
'ip link show', but this output format is not supported on the
implementation of ip present on OpenWrt devices. This change updates
LinkWithPrefix to use the full format of 'ip link show' instead
of the brief format. Usage and expected output remain the same. Unit
tests have also been added for the new version of LinkWithPrefix.

BUG=b:211517577
TEST=tast unit tests passed

Change-Id: Ifce58e3c9841092adb67bf773b07301005bd165a