tast-tests: Refactor bridge and veth cleanup into common
This change refactors the bridge and veth cleanup steps preformed on
router start and close into the common package so that they may be
reused in the OpenWrt router implementation. Specific cleanup
functions for bridge and veth have been added as well to make the
intent during cleanup more clear and consistent.

BUG=b:211517577
TEST=Successfully ran tast test wifi.RoamFT.psk, which uses the
bridge and veth support, on a wificell with a legacy router.

Change-Id: I0bf8efc853617b45e593cdeb9116605468fcc065
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3651561
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>