crostini: Add a new test for the CrostiniUseLxd4 feature
We've had multiple issues now from crostini failing to properly
respect this feature, so add a test for it.

This test just installs crostini with the feature enabled and checks
that the expected version of LXD is running. This does not attempt to
verify that updating an existing VM will work.

BUG=chromium:1097610, chromium:1321396
TEST=Ran test

Change-Id: I355821af2491d6e1fed893473fd2b17248466cd3