autoupdate: Add auto_update_stable dependency to tests
Autoupdate tests fail if the board does not have available
updates. The flag auto_update_stable allows filtering out most
of those boards to avoid running into this expected failure.

Update EnterpriseRollbackPreviousVersion to fail if an available
image is not found instead of logging a message, as now the image
should be found.

BUG=None
TEST=tast run <DUT> autoupdate.*

Change-Id: If80522d77fffb2b89f368ab2dbb17197a683c966