tast-tests: fix VK lock screen test
inputs.VirtualKeyboardLockScreen leaves the DUT in the lock screen
upon completion, and the fixture cannot reset to logged in state for
reuse. Thus the following tests all fail.

Remove the fixtures from the test, and instead create new Chrome,
will revisit this after the fixtures that restart the Chrome session are
available.

BUG=None
TEST=tast run <DUT> inputs.VirtualKeyboardLockScreen*

Change-Id: I72a7fd8ee08b7e48863c14353c97f92081a474af