usbprinter: Ensure debugd is running when adding printers
Chrome uses debugd to set up the newly plugged-in printer when
virtual-usb-printer starts.  In case an earlier test causes debugd to
exit, tast normally tries to start debugd before each test, but it
doesn't treat it as a fatal error if it fails.  Since printing depends
directly on debugd, make sure it's started whenever we launch the
virtual-usb-printer.

BUG=b:230787720
TEST=700 cycles of tast run printer.{Print,PrintIPPUSB,PrintUSB}

Change-Id: I365d4ef59e98295d8c404a81580a0f9251bb4c4f