tast-tests: WiFi: Update the calculation of dwellTime in wifi.ChannelScanDwellTime
Probe requests from the DUT may not be captured in the pcap file as the
AP keeps sending beacon frames, which fails the test.

Update the calculation of dwellTime in wifi.ChannelScanDwellTime so that
the start time is the timestamp of the first beacon frame returned by
the active scan. This doesn't require the probe request, but introduces
an error of no more than 1ms (the DUT sends a probe request immediately after receiving a beacon frame, and the interval of two beacons is 1ms).

BUG=b:225073589, b:231669774
TEST=tast run -var=router=localhost:2210 -var=pcap=localhost:2220
localhost:2200  wifi.ChannelScanDwellTime

Change-Id: Ie8f303ad8a75d71971161bcb799e812b0f0fafd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3575213
Tested-by: Kaidong Wang <kaidong@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Auto-Submit: Kaidong Wang <kaidong@google.com>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Commit-Queue: Kaidong Wang <kaidong@google.com>