tast-tests: stabilize crostini.AppGeditInstallUninstall.* tests
crostini.AppGeditInstallUninstall.* tests fail on some devices, because
as a part of the test we open a "save as" dialog and attempt to press
"save" in it. This fails, if the "save" button from the main editor
window is visible.

Scope ACUITI to look for a "save" button within "save as" window only.

BUG=b:232980375
TEST=tast run dut crostini.AppGeditInstallUninstall.stable

Change-Id: I2256a7e186ab88dddfeed1284dd892fe4b7431b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3647620
Tested-by: Alexey Shpakov <ashpakov@google.com>
Commit-Queue: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Alvin Jia <alvinjia@google.com>