tast-tests: stabilize crostini fixtures
crostini fixtures could sometimes need to uninstall crostini. As part
of removal we continuously left click delete button in a confirmation
dialog until RemoveLinuxAlert appears. For that, currently we use
WaitUntilExists as a condition to stop clicking. This method polls for
a condition until it is fulfilled. This works, if the first left click
succeeded. However, if it did not, WaitUntilExists will poll until ctx
is exhausted before returning. As the ctx duration it uses matches that
of the left click, once it returns due to a timeout, we no longer have
time left for yet another left click.

Replace double-polling introduced by WaitUntilExists with a one off check
to allow for repeated left clicks.

BUG=b:231630463
TEST=tast run dut crostini.AppGeditInstallUninstall.buster_stable

Change-Id: Ia01a4b640d084cff6ffefdce9a4dd1eb82e924b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3653780
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Commit-Queue: Alexey Shpakov <ashpakov@google.com>