Lacros: Simplify initialization routine in StartupPerf
This CL simplifies the main initialization routine of StartupPerf by
moving it out of the iteration loop.

There is no need to call performInitialLogin fn every time before
actually measuring and, besides, doing it out of the loop will make the
whole testing faster on Crosbolt.

In fact, running performInitialLogin out of the iteration loop is closer
to the expected CUJ behavior here (because some user states are not
cleaned during sign out -> sign in -> waiting for Lacros to be restored).
That said, it's rather important to measure cold boot startup although
that's not the intention of StartupPerf.

BUG=chromium:1270977
TEST=tast run $dut lacros.StartupPerf.rootfs_primary

Change-Id: I919391d27830b78240689026cdfd431437b7be1e