xmlrpc: Remove the max timeout
There is no reason to limit the timeout of xmlrpc calls. If someone
passes in a long deadline, use it.

Since switching from dd to the servo control `download_image_to_usb_dev`
usb key flashing has been failing because 90s is not long enough for the
servo to flash the USB stick.

BUG=b:224586052
TEST=None

Change-Id: Ice92edc6dc63bedea2f8e177473392fdcbcab6c1