Setup Captive Portal Ethernet Properties Test
This test checks the manager and service properties related to captive
portal and verifies that they are what we expect the properties to be.
The test uses the already online ethernet technology to verify the
properties. This will also give future captive portal tests a baseline
on what properties to test against.

BUG=b:233077037
TEST=tast run ${DUT} network.ShillCaptivePortalPropertiesEthernet

Change-Id: I4969397816c8989e14ce3e5870370bacc716ce35