Setup Captive Portal Ethernet Properties Test
This test is a baseline test for captive portal and verifies that the state property in the service is online, which is what we expect it to be. This will also give future captive portal tests a very basic baseline test. This is a precursor to the implementation of Virtual Ethernet for the tests.

BUG=b:233077037
TEST=tast run ${DUT} network.ShillCaptivePortalProperties

Change-Id: I4969397816c8989e14ce3e5870370bacc716ce35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3653012
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kartik Hegde <khegde@chromium.org>
Tested-by: Michael Rygiel <michaelrygiel@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Michael Rygiel <michaelrygiel@google.com>
Commit-Queue: Michael Rygiel <michaelrygiel@google.com>