audio.UCMSequences: Use fixture to reduce sleep time
audio.UCMSequences.* tests require CRAS to be not running
for at least 11 seconds to ensure exclusive access to audio
devices.

Those 3 subtests can share the "CRAS is stopped" setup, and when
ran consecutively do not need sleeps in between.
Introduce and use a "crasStopped" fixture so the subtests
only sleep for 11 seconds, instead of 3*11 seconds.

BUG=b:229558276
BUG=b:231671430
TEST=Run the test when playing a video on youtube:
     tast run $DUT audio.UCMSequences.\*

Change-Id: I330bd7f2ab0581316ed3e948e9f845a131154ea2