[tast-tests] Verify shelf app movement during rearrangement
This CL moves a shelf app button from the last pinned app slot to the
middle of shelf. Verify that the apps that are not under drag are moved
as expected.

In detail, this CL does the following things:

(1) Change some variables' names in order to improve code readability.
(2) Ensure that there are four pinned apps for each test type so that
the test context becomes more complex.
(3) Besides moving an app from the last shelf app slot to the first slot
directly, hover the dragged app in the middle part of the shelf. It
helps to verify that those pinned apps that are not under drag have the
expected movement.

BUG=b:200987018
TEST=shelf.AppRearrangement.*

Change-Id: I906b036aafd491e81e55d6f3c4bb4f4738b7026a