network: vpn: Fix routing for wireguard server
This patch adds the route to route back packet correctly in the netns
for wireguard. Currently we only test the VPN server is pingable,
without this route, ping reply packets will go via the veth interface
instead of the wg interface which is wrong. This wrong behavior does not
affect the current VPNConnect tests because the host side can still get
this packet from the veth interface and thus ping is happy.

BUG=None
TEST=tast run $DUT network.VPNConnect.wireguard*

Change-Id: I53c4318df20cd019f3ce59c897d185574e40bb42