health: Log in Chrome and disable data protection in fixture
When loggin in Chrome, cros_healthd will restart. Once our fixture
detects PID change, it'll mark the test as fail.

In CL:3471438, the owner modified the fixture configuration so
"thunderbolt" subtest won't run the healthd fixture. But this is not
okay, so in CL:3635402, we fixed it. This cause the b/233137825 issue.

After some discussion, we decided to move the logic into fixture part to
do the Chrome log in. It means that all healthd Tast tests are run after
log in, but it should be fine because we should still work well in this
case.

BUG=b:233137825
TEST=1. tast run $dut health.ProbeBusInfo.*
     2. Tests can pass without cros_healthd PID change.

Change-Id: I292d1e7b83b9628af3f4d0464ef1dff2fb7f3fee