familylink: Add lacros variant for unicorn login tast test
Add a lacros variant for the unicorn login tast test. This change also
allows easily creating lacros variants for any familylink fixture. It also
changes how the signed-in user is detected, because the “account”
row on the family link user internals page is dependent on sync being
enabled, and it is not trivial to enable sync in lacros for a tast test.

BUG=chromium:1313067
TEST=both variants of unicorn login pass

Change-Id: I3e61196e4eff27cb141b84e774f9d497fabe184a