familylink: Add Lacros variant for Unicorn login Tast test
Add a Lacros variant for the Unicorn login Tast test. This change also
allows easily creating Lacros variants for any familylink fixture. It
also changes how the signed-in user is detected, because the “account”
row on the family link user internals page is dependent on sync being
enabled, and it is not trivial to enable sync in Lacros for a Tast test.
This implementation of VerifyUserSignedIntoBrowserAsChild works in both
ash and Lacros.

BUG=chromium:1313067
TEST=both variants of unicorn login pass

Change-Id: I3e61196e4eff27cb141b84e774f9d497fabe184a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3655634
Tested-by: Galen Corey <galenemco@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Toby Huang <tobyhuang@chromium.org>
Commit-Queue: Galen Corey <galenemco@chromium.org>