tast-tests: Use log to dump processes.
Currently, there seems some issue on dump output files.
We're investigating the issue, but the code edited by this CL is
also to help investigation of another test flakiness.
In order to run those investigation in parallel, this CL starts
to output process dump to log lines, which is not recommended
in general, though.

BUG=b:231683759
TEST=./fast_build.sh -T

Change-Id: I74f30adf9d066b46f665d6b223c71223e6b656c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3655414
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>