arc: Skip pin setup screens if shown
After changes to when we show the PIN setup screen during OOBE (see
linked bug), arc.OobeArc and arc.OobeArcAppOpen started failing more
often. This is because these tests don't properly skip the PIN setup
screen if it is shown. This commit patches the two tests so that they
skip the PIN screen if it is shown. See also oobe.gestureNavigation,
which already did this.

BUG=chromium:1308729
TEST=tast -verbose run <DUT> arc.OobeArc.*
TEST=tast -verbose run <DUT> arc.OobeArcAppOpen.*

Change-Id: I5b06ad7e34d10da95082fddc6ddb95eabc09779a