[tast-tests] Verify shelf app rearrangement under RTL
This CL verifies shelf app rearrangement under RTL. In detail, this CL
makes the current Tast test for shelf app rearrangement parametized. To
ensure the system UI under RTL, this CL sets the language to be Arabic.

This CL verifies app rearrangement by dragging a Chrome extension and a
File app around the shelf. The verification with a PWA app and an
Android app will be implemented in the subsequent CLs (maybe in low
priority since it is not requested by the test procedure).

BUG=b:200987018
TEST=shelf.AppRearrangement.*

Change-Id: I054d6af365d06c53d94fecf40d6ec8e03b8563a7