tast-tests: stabilize crostini fixtures
crostini fixtures setup has a separate step to uninstall crostini
if it happens to be installed at the start of the fixture. During
that step we click a button to uninstall crostini and wait for a
confirmation dialog. In some cases the click doesn't register and
we end up never getting the confirmation dialog.

Retry clicking multiple times until confirmation dialog appears.

BUG=b:233813429
TEST=tast run dut crostini.AppGeditInstallUninstall.stable

Change-Id: I9b41661c5d65527d01c7b61cd044d7669fd3d5ab