tast-tests: Fix setting “solid” collection on small screen
The solid collection is at the end of the collection list so when
the screen height is too small, the wallpaper isn’t visible.
We need to scroll down to make it visible.

We already have “scroll down” logic for selecting the _wallpaper_,
so this CL just extracts the logic into scrollDownUntilSucceeds()
and reuses it for selecting the _collection_.

BUG=b:232865473
TEST=tast run arc.ResizeLock arc.CompatSnap
TEST=tast unit tests passed.

Change-Id: Ib76b0d4ab0aa9e57e96cac9073fc89cee41e703e