tast-tests: stabilize crostini fixtures
As part of crostini installation we continuously click turn on and next
buttons until other buttons appear. For that, currently we use
WaitUntilExists as a condition to stop clicking. This method polls for
a condition until it is fulfilled. This works, if the first left click
succeeded. However, if it did not, WaitUntilExists will poll until ctx
is exhausted before returning. As the ctx duration it uses matches that
of the left click, once it returns due to a timeout, we no longer have
time left for yet another left click.

Replace double-polling introduced by WaitUntilExists with a one off
check to allow for repeated left clicks.

BUG=b:231630463
TEST=tast run dut crostini.AppGeditInstallUninstall.stable

Change-Id: I123dfa17ae319afae74f456caaa1e9a8d45f0aea