arc/gio: Force input overlay test application to be windowed mode
For some reason, the input overlay test application is launched as
default as fullscreen on all ARC-P builds; this causes the
"Activity.SurfaceBounds" tast function to fail, since it does not take into account the caption height. In order to ensure test
uniformity and correctness, the application will be forced to launch in
windowed mode.

BUG=b:229156145
TEST=./fast_build.sh -T && tast run $DUT arc.InputOverlay*
TEST=tested on ARC-P devices

Change-Id: I211e0a2339282f2121290f6478934f48a49db2c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3669553
Commit-Queue: Joseph Pang <pjlee@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Joseph Pang <pjlee@google.com>
Reviewed-by: Cici Ruan <cuicuiruan@google.com>