tast-tests: add un/stable variant for holdingspace.Download
This is a prep work for promoting lacros_pin_and_unpin.

The test is split to one for lacros_stable (without VMs)
and the other for lacros_unstable (with VMs) to be able to promote
a stable variant in a follow up.

This also replaces /home/chronos/user path that is being deprecated
with cryptohome.DownloadsPath.

BUG=chromium:1303875
TEST=tast run ${DUT} holdingspace.Download*

Change-Id: Ib300d1c938ff7891df55af3f51176aa737388d9f