[FilesTrash] Add filemanager.SendToTrash test for Drive and local
Start verifying whether files can be trashed in both the DriveFS and
local disk volumes. Exclude the crostini root for now until we have
landed support in the IOTask for this. This is to ensure as we
transition from background js to IOTask the feature continues to work.

BUG=b:231830484
TEST=tast run filemanager.SendToTrash

Change-Id: Ie93cf5d9d2c0f12c3530500e16ff634fed0ace02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3670618
Reviewed-by: Alvin Jia <alvinjia@google.com>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>