tast-tests: stabilize crostini.DragDrop.*
1) crostini.DragDrop.* tests are not written with tablets and
detachables in mind. As result, whenever it is run on a tablet in a
portrait mode it fails, because it needs to arrange two windows side
by side for drag and drop. This is not supported in portrait mode.
2) Once this is fixed, we uncover another issue: drag and drop in the
direction of the Files app accidentally lands on downloads directory.

1) Check orientation of the device at the start of the test and switch
to landscape mode, if needed.
2) Drag closer to the bottom of the applet to avoid accidental
collisions with existing files and directories in the Files app.

BUG=b:231512029
TEST=(druwl,volteer) tast run dut crostini.DragDrop.bullseye_stable

Change-Id: Iaf199493dcc8a060a9dc7ff97aaee6f3608ff443