ui: Improve consistency of WindowArrangementCUJ testing
ui.WindowArrangementCUJ.tablet_mode* will arrange for an empty overview
grid by dragging the overview window to a second desk instead of closing
the overview window. The problem with closing the overview window is
that an associated PIP window is then also closed but we would prefer to
keep that PIP window open.

BUG=b:222716317
TEST=tast run <DUT> ui.WindowArrangementCUJ.tablet_mode*

Change-Id: I0a4fb0c12b148442d851472fbaf6be171ea0ef7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3668699
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Tested-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Joseph Pang <pjlee@google.com>