typec: Make Basic a critical test
There is still the occasional failure, but the pass rates are high
enough that there is good value in the coverage provided by typec.Basic,
so make it critical.

There are some boards on which the test has occasional failures, so move
those into an "unstable" version of the test, which is marked
informational.

BUG=b:234056251
TEST=Passing 10/10 times on brya and strongbaed.

Change-Id: I168b3b8976bfc3444df51f59c8870a8422b6cb85