tast: Make emoji suggestion tests more robust.
Currently the emoji suggestion tests look for a StyledLabel with the
correct emoji text. However, we want to refactor the UI so that there's
no StyledLabel.

Just take the first element that has the correct emoji text, since the
emoji is pretty unique and clicking any element with the emoji text
will trigger emoji submission.

BUG=b:233264555
TEST=tast run dut inputs.PhysicalKeyboardEmojiSuggestion*

Change-Id: I477a6cc1a5d7d2e5a7e83d1cede4321c21ba8a9e