Fix up removeMount SMB call
The chrome.fileManagerPrivate.removeMount call changed to include a no
argument callback to indicate success. I made a mistake in the CL review
assuming the tast.promisify function was available when it isn't in the
FIles app context (it's only on the test connection). Fix that by
instead wrapping the fileManagerPrivate API call correctly.

BUG=None
TEST=tast run filemanager.SMBFileOperation*

Change-Id: I9cc2975037c2b2d76a71979a937c7a6429522d99