tast-tests: Do not wait for package restrictions to be set
CloudDPC for ARC has been updated to set the package restrictions before
the package is installed so we don't need to wait for the package
restrictions to be set when the app is already installed since setting
the restrictions is much quicker local operation than downloading an app
from internet.

See http://cl/451322257 for details.

BUG=b:231791924
TEST=tast run DUT enterprise.ARCProvisioning.vm_unstable

Change-Id: If401dc57951ac98348c5c2e96c7838da8289db9e