tast-tests: Wait for pacakge to unfreeze before considering ready
enterpirse.ARCInstallLogging test fails often due to launch attempt
being made while the app package is frozen. This occurs because package
manager publishes the package in the list of installed packages before
unfreezing the package and we launch the app soon after we detect that
the package is installed so there is a race that results in app being
launched before package manager has compeleted the install flow.

In this change, we're adding a new helper method to arc package to list
frozen packages in the package manager using dumpsys and using that list
to determine if the package is installed and ready for use.

BUG=b:228857862
TEST=tast run DUT enterprise.ARCInstallLogging.vm

Change-Id: Ib003f68f9c0c78cc449121d96bc2ef1529bae792