tast-tests: stabilize crostini fixture
Crostini installation fixture uses LeftClickUntil in multiple places.
Here are three potential options for the until condition considered
and their behaviours in case of a flake:
(1) ui.LeftClickUntil(<node1>, ui.WaitUntilExists(<node2>))
 - click node1
 - wait until node2 exists or timeout hits
(2) ui.LeftClickUntil(<node1>,
  ui.WithTimeout(shortUITimeout).WaitUntilExists(<node2>))
 - click node1
 - wait until node2 exists or shortUITimeout hits
 - click node1
 - wait until node2 exists or shortUITimeout hits
 - repeat until timeout hits
(3) ui.LeftClickUntil(<node1>, ui.Exists(<node2>))
 - click node1
 - check that node2 exists
 - click node1
 - check that node2 exists
 - repeat until timeout hits
(4) ui.WithInterval(shortUITimeout).LeftClickUntil(<node1>,
 ui.Exists(<node2>))
 - click node1
 - sleep for shortUITimeout
 - check that node2 exists
 - click node1
 - sleep for shortUITimeout
 - check that node2 exists
 - repeat until timeout hits

(1) works well if we need to wait until timeout for a codition,
but it fails, if the first click did not register.
(2) waits less than (1) for a condition, but repeats a click
in-between - this will help in case a click hasn't registered.
(3) keeps clicking every 500ms and checking for a condition -
it ends up spamming clicks and a slow board may not have enough time
to react. It caused b:234128639 .
(4) is very similar to (2), the difference is that instead of
continuously polling for node2 presence, it sleeps for interval
duration before checking for presence once. That is, it will always
take shortUITimeout interval to complete, even if the action suceeded
at once.

Adjust crostini fixture to use approach (2) in its interactions.

BUG=b:231630463
TEST=tast run dut crostini.Basic.bullseye_stable

Change-Id: I4f59cc790ddb1239d75035292b6bdcd3b2bef4ba