crostini: Capture startup perf data in fixtures
In Chrome we have UMA metrics for the CrostiniRestarter.
This is used by the installer and any subsequent crostini
launches, so we sample the before and after values of
the Crostini.RestarterTimeInState2 histograms before and
after installing Crostini in the fixture.

BUG=b:231511895
TEST=tast run <DUT> crostini.xxx, check output directory for
results.json file.

Change-Id: I32c100d6fd8b8c5c2bae406cde05475d42636014
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3676401
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Nicholas Verne <nverne@chromium.org>
Tested-by: Nicholas Verne <nverne@chromium.org>