power: Use pointers for "BatteryDischargeMode"
Change to use pointer receivers for the structs that implement
BatteryDischargeMode interface, and return pointers from the functions
that return BatteryDischargeMode.
This change fixes the problem caused by passing BatteryDischargeMode
value in powersetup.PowerTestOptions that the actual discharge error
could not be returned with Err() function in the caller function.

BUG=b:234429108
TEST=./fast_build.sh -T

Change-Id: I827ba6c0ec2be7e6b611411efb93ba6315b5e8ac