faft: Improve on ScreenWakeTabletMode
Close rpc connection before closing lid, and reconnect
after lid opened. This might help with the context deadline
error in taking a screenshot. Also, increment the duration
of power press to turn display off when retries fail, until
a total of 1 second is reached. Lastly, expand on the list
of convertibles with keyboard scanned in tablet mode, based
on recent Stainless results.

BUG=b:200305356
TEST=manual on Eldrid, Asurada, Storo360, and Delbin

Change-Id: I85e95e521e8fa3c6a5a943bab1452dc120dc8f79