tast-tests: Test networks config during oobe
Add method to netconfig that creates a connection to
cros_network_config during OOBE.

Add example test to check the API works as expected and it is possible
to make calls to cros_network_config.mojom. It sets a simple
configuration and tests it then exists as expected. This sample test
is only for reference so it is set as manual execution.

BUG=b:229103041, b:207087735
TEST=tast -verbose run DUT network.TestConfigDuringOobe

Change-Id: Ifd141e929e4df9b9c60a5521d5225b3d83878d0c